split sdma-ep into device, host, tester header files
